refactor: remove duplicate initialization code in profiles
