+++ examples/basic.html
@@ -3,7 +3,7 @@
   <head>
     <meta charset="utf-8" />
     <title>Replace me with a real title</title>
-    <script src="../profiles/w3c.js" async class="remove"></script>
+    <script src="../profiles/w3c.js" type="module" class="remove"></script>
     <script class="remove">
       // All config options at https://respec.org/docs/
       var respecConfig = {
