+++ examples/example_aom.html
@@ -5,7 +5,7 @@
     <title>
       Replace me with a real title
     </title>
-    <script src='../profiles/aom.js' async class='remove'></script>
+    <script src='../profiles/aom.js' type="module" class='remove'></script>
     <script class='remove'>
       var respecConfig = {
         specStatus: "PD",
