+++ profiles/aom.js
@@ -1,13 +1,7 @@
-"use strict";
-// In case everything else fails, we want the error
-window.addEventListener("error", ev => {
-  console.error(ev.error, ev.message, ev);
-});
+import * as ReSpec from "../src/respec.js";
 
 const modules = [
   // order is significant
-  import("../src/core/base-runner.js"),
-  import("../src/core/ui.js"),
   import("../src/core/location-hash.js"),
   import("../src/core/l10n.js"),
   import("../src/aom/defaults.js"),
@@ -59,23 +53,6 @@ const modules = [
   import("../src/core/linter-rules/no-http-props.js"),
 ];
 
-async function domReady() {
-  if (document.readyState === "loading") {
-    await new Promise(resolve =>
-      document.addEventListener("DOMContentLoaded", resolve)
-    );
-  }
-}
-
-(async () => {
-  const [runner, { ui }, ...plugins] = await Promise.all(modules);
-  try {
-    ui.show();
-    await domReady();
-    await runner.runAll(plugins);
-  } finally {
-    ui.enable();
-  }
-})().catch(err => {
-  console.error(err);
-});
+Promise.all(modules)
+  .then(plugins => ReSpec.run(plugins))
+  .catch(err => console.error(err));
